8222286: Fix for JDK-8213419 is broken on s390 Summary: Cast sizeof argument to log2_intptr to uintptr_t to remove ambiguity Reviewed-by: aph, clanger
(cherry picked from commit 4a4ca84e395dd0879c6f0a7046d9045b0bd20dbf)

+++ src/hotspot/share/runtime/os.cpp
@@ -1381,7 +1381,8 @@ char** os::split_path(const char* path, int* n) {
 }
 
 void os::set_memory_serialize_page(address page) {
-  int count = log2_intptr(sizeof(class JavaThread)) - log2_int(64);
+  // S/390: Cast to (uintptr_t) to remove ambiguity; see JDK-8222286
+  int count = log2_intptr((uintptr_t) sizeof(class JavaThread)) - log2_int(64);
   _mem_serialize_page = (volatile int32_t *)page;
   // We initialize the serialization page shift count here
   // We assume a cache line size of 64 bytes
